newMovie refactored
