+++ src/App.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import './App.scss';
 import { MoviesList } from './components/MoviesList';
 import { NewMovie } from './components/NewMovie';
-import { Movie } from './types/Movie';
+import { Movie } from './types/typedefs';
 import moviesFromServer from './api/movies.json';
 
 export const App: React.FC = () => {
