+++ src/components/NewMovie/NewMovie.scss
