--- src/types/Movie.ts
@@ -1,7 +1,0 @@
-export interface Movie {
-  title: string;
-  description: string;
-  imgUrl: string;
-  imdbUrl: string;
-  imdbId: string;
-}
